refactor: MyType 폴더 구조 재조직 #43
