--- lib/type/component/component.dart
@@ -1,2 +1,0 @@
-export 'package:typed/type/component/add_record_dialog.dart';
-export 'package:typed/type/component/grid_item_widget.dart';
