+++ lib/type/models/grid_item.dart
