+++ lib/type/models/split_view_state.dart
