--- lib/type/presentation/models/index.dart
@@ -1,1 +1,0 @@
-export 'package:typed/type/presentation/models/grid_item.dart';
