--- lib/type/screen/screen.dart
@@ -1,5 +1,0 @@
-export 'package:typed/type/screen/my_type.dart';
-export 'package:typed/type/screen/my_sentence_record_screen.dart';
-export 'package:typed/type/screen/my_book_record_screen.dart';
-export 'package:typed/type/screen/my_music_record_screen.dart';
-export 'package:typed/type/screen/my_image_record_screen.dart';
