+++ lib/type/screens/component/add_record_dialog.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:typed/common/const/index.dart';
+import 'package:typed/type/screens/screen.dart';
+import 'package:typed/type/models/grid_item.dart';
 import 'package:image_picker/image_picker.dart';
-import 'package:typed/common/const/app_themes.dart';
-import 'package:typed/type/presentation/models/index.dart';
-import 'package:typed/type/screen/screen.dart';
 
 class AddRecordDialog extends StatelessWidget {
   final XFile? initialImage;
