+++ lib/type/screens/component/empty_record_dialog.dart
