+++ lib/type/screens/my_image_record_screen.dart
@@ -1,12 +1,11 @@
-import 'dart:io';
 import 'package:flutter/material.dart';
+import 'dart:io';
+import 'package:typed/common/const/index.dart';
+import 'package:typed/common/index.dart';
+import 'package:typed/type/models/grid_item.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:permission_handler/permission_handler.dart';
-import 'package:typed/common/const/app_bar_style.dart';
-import 'package:typed/common/const/app_themes.dart';
-import 'package:typed/common/index.dart';
-import 'package:typed/type/presentation/models/index.dart';
 
 class MyImageRecordScreen extends ConsumerStatefulWidget {
   final XFile? initialImage;
