+++ lib/type/screens/my_sentence_record_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:typed/type/layout/my_type_layout.dart';
+import 'package:typed/type/screens/layout/my_type_layout.dart';
 
 class MySentenceRecordScreen extends StatefulWidget {
   const MySentenceRecordScreen({super.key});
