+++ lib/type/screens/screen.dart
@@ -1,0 +1,5 @@
+export 'package:typed/type/screens/my_type.dart';
+export 'package:typed/type/screens/my_sentence_record_screen.dart';
+export 'package:typed/type/screens/my_book_record_screen.dart';
+export 'package:typed/type/screens/my_music_record_screen.dart';
+export 'package:typed/type/screens/my_image_record_screen.dart';
